Separated instructions into types
