+++ src/Instructions/CMakeLists.txt
@@ -2,20 +2,25 @@ cmake_minimum_required(VERSION 3.16)
 
 # Set include directory
 include_directories("${PROJECT_SOURCE_DIR}/include/Instructions")
+include_directories("${PROJECT_SOURCE_DIR}/include/Instructions/Arithmetic")
+include_directories("${PROJECT_SOURCE_DIR}/include/Instructions/Comparison")
+include_directories("${PROJECT_SOURCE_DIR}/include/Instructions/FlowCtrl")
+include_directories("${PROJECT_SOURCE_DIR}/include/Instructions/Logic")
+include_directories("${PROJECT_SOURCE_DIR}/include/Instructions/Other")
 
 # Set source files
 set(SOURCES
-    ADDInstruction.cpp
-    ANDInstruction.cpp
-    BInstruction.cpp
-    BICInstruction.cpp
-    CMNInstruction.cpp
-    CMPInstruction.cpp
-    EORInstruction.cpp
+    Arithmetic/ADDInstruction.cpp
+    Logic/ANDInstruction.cpp
+    FlowCtrl/BInstruction.cpp
+    Logic/BICInstruction.cpp
+    Comparison/CMNInstruction.cpp
+    Comparison/CMPInstruction.cpp
+    Logic/EORInstruction.cpp
     InstructionBase.cpp
     InstructionRepository.cpp
-    MOVInstruction.cpp
-    ORRInstruction.cpp
+    Other/MOVInstruction.cpp
+    Logic/ORRInstruction.cpp
 )
 
 # Make libraries
