+++ src/Instructions/Comparison/CMNInstruction.cpp
