+++ src/Instructions/Logic/BICInstruction.cpp
